+++ lib/app_constants.dart
@@ -1,0 +1,3 @@
+final String appName = 'Zengage';
+
+final String homePageTitle = 'Welcome';
